Updated: 구조 개선 [컨트롤러 스웨거 의존성 분리]
